chore(auth): Refactor to eliminate circular dependencies
