+++ packages/auth/src/providers/cognito/apis/signInWithCustomAuth.ts
@@ -9,11 +9,9 @@ import { assertValidationError } from '../../../errors/utils/assertValidationErr
 import { assertServiceError } from '../../../errors/utils/assertServiceError';
 import {
 	getActiveSignInUsername,
-	getNewDeviceMetadata,
 	getSignInResult,
 	getSignInResultFromError,
 	handleCustomAuthFlowWithoutSRP,
-	retryOnResourceNotFoundException,
 } from '../utils/signInHelpers';
 import { InitiateAuthException } from '../types/errors';
 import {
@@ -32,6 +30,8 @@ import {
 } from '../../../foundation/factories/serviceClients/cognitoIdentityProvider/types';
 import { tokenOrchestrator } from '../tokenProvider';
 import { dispatchSignedInHubEvent } from '../utils/dispatchSignedInHubEvent';
+import { retryOnResourceNotFoundException } from '../utils/retryOnResourceNotFoundException';
+import { getNewDeviceMetadata } from '../utils/getNewDeviceMetadata';
 
 /**
  * Signs a user in using a custom authentication flow without password
